Перенос «ActionCreator» в отдельную директорию.
